# university/src/main/resources/openapi.v1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timetable",
    "version": "1.0",
    "contact": {
      "name": "Taras Shablii"
    },
    "description": "Timetable App API spec"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1/timetable"
    }
  ],
  "paths": {
    "/students": {
      "get": {
        "summary": "List all students",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "operationId": "get-students",
        "description": "get all students (optionally paginated)",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "in": "query",
            "name": "page"
          }
        ]
      },
      "parameters": [],
      "post": {
        "summary": "Create new student",
        "operationId": "post-students",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "create new student",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        }
      }
    },
    "/students/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Retrieve student",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-students-studentId",
        "description": "get single student by id",
        "parameters": []
      },
      "put": {
        "summary": "Edit existing student",
        "operationId": "put-students-id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "description": "edit student",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          },
          "description": ""
        }
      },
      "delete": {
        "summary": "Delete student",
        "operationId": "delete-students-id",
        "responses": {
          "501": {
            "description": "Not Implemented"
          }
        },
        "description": "delete student"
      }
    },
...
    "/groups/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Retrieve group",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-groups-id",
        "description": "get single group by id"
      },
      "put": {
        "summary": "Edit existing group",
        "operationId": "put-groups-id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "edit group",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Group"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete group",
        "operationId": "delete-groups-id",
        "responses": {
          "501": {
            "description": "Not Implemented"
          }
        },
        "description": "delete group"
      }
    },
...
    "/professors/{id}/courses/{courseId}/students": {
      "parameters": [
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "id",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": "courseId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "List all students attending professor's course",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "operationId": "get-professors-id-courses-courseId-students",
        "description": "get students attending professor's course"
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "title": "Student",
        "type": "object",
        "description": "",
        "x-examples": {},
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "exclusiveMinimum": false
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "group": {
            "$ref": "#/components/schemas/Group"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
...
      "Group": {
        "title": "Group",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {}
  }
}